zip: Decode Unix extra-field timestamps as signed
atime/mtime/ctime in the 0x000D (PKWARE Unix), 0x5455 (Info-Zip "UT")
and 0x5855 (Info-Zip "UX", old) extra fields are specified as signed
32-bit Unix time, to allow pre-1970 dates. archive_le32dec() returns an
unsigned uint32_t though, and its result was being assigned directly to
zip_entry->{a,m,c}time (a signed time_t), so any negative timestamp got
reinterpreted as a huge positive one instead - wrapping forward by 2^32
seconds (about 136 years).

Cast through int32_t at each of these 7 call sites so the sign is
preserved when the value widens into time_t.

Reproduced with bsdtar alone (both sides are libarchive):

  $ touch -d '1969-12-31 23:59:58 UTC' old.txt
  $ bsdtar -cf rt.zip --format=zip old.txt
  $ bsdtar -tvf rt.zip
  -rw-rw-r--  0 1001   1001        0 Feb  7  2106 old.txt

Added regression tests for negative Unix extra-field timestamps.
Verified these fail without the fix: reverting the signed cast wraps
-86400 to 4294880896, matching the ~2106 wraparound described in the bug
report.

Bug: https://github.com/libarchive/libarchive/issues/3550

+++ libarchive/archive_read_support_format_zip.c
@@ -762,9 +762,12 @@ process_extra(struct archive_read *a, struct archive_entry *entry,
 		case 0x000d:
 			/* PKWARE Unix Extra Field fixed metadata. */
 			if (datasize >= 12) {
-				zip_entry->atime = archive_le32dec(p + offset);
+				/* atime/mtime are signed 32-bit Unix
+				 * time, to allow pre-1970 dates. */
+				zip_entry->atime =
+				    (int32_t)archive_le32dec(p + offset);
 				zip_entry->mtime =
-				    archive_le32dec(p + offset + 4);
+				    (int32_t)archive_le32dec(p + offset + 4);
 				zip_entry->uid =
 				    archive_le16dec(p + offset + 8);
 				zip_entry->gid =
@@ -815,7 +818,10 @@ process_extra(struct archive_read *a, struct archive_entry *entry,
 #endif
 				if (datasize < 4)
 					break;
-				zip_entry->mtime = archive_le32dec(p + offset);
+				/* Signed 32-bit Unix time, to allow
+				 * pre-1970 dates. */
+				zip_entry->mtime =
+				    (int32_t)archive_le32dec(p + offset);
 				offset += 4;
 				datasize -= 4;
 			}
@@ -823,7 +829,8 @@ process_extra(struct archive_read *a, struct archive_entry *entry,
 			{
 				if (datasize < 4)
 					break;
-				zip_entry->atime = archive_le32dec(p + offset);
+				zip_entry->atime =
+				    (int32_t)archive_le32dec(p + offset);
 				offset += 4;
 				datasize -= 4;
 			}
@@ -831,7 +838,8 @@ process_extra(struct archive_read *a, struct archive_entry *entry,
 			{
 				if (datasize < 4)
 					break;
-				zip_entry->ctime = archive_le32dec(p + offset);
+				zip_entry->ctime =
+				    (int32_t)archive_le32dec(p + offset);
 				offset += 4;
 				datasize -= 4;
 			}
@@ -841,9 +849,12 @@ process_extra(struct archive_read *a, struct archive_entry *entry,
 		{
 			/* Info-ZIP Unix Extra Field (old version) "UX". */
 			if (datasize >= 8) {
-				zip_entry->atime = archive_le32dec(p + offset);
+				/* Signed 32-bit Unix time, to allow
+				 * pre-1970 dates. */
+				zip_entry->atime =
+				    (int32_t)archive_le32dec(p + offset);
 				zip_entry->mtime =
-				    archive_le32dec(p + offset + 4);
+				    (int32_t)archive_le32dec(p + offset + 4);
 			}
 			if (datasize >= 12) {
 				zip_entry->uid =
